api_test/requirements.txt
Django==3.2.6
celery==5.2.3
redis==3.5.3
djoser==2.1.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.8.0